Further improvements to M5Stack Cardputer keyboard integration.
 - Move the keyboard-related code from board.c to
   cardputer_keyboard.c, exposed to python as cardputer_keyboard.
 - Remove board.KEYBOARD, provide cardputer_keyboard.KEYBOARD instead.
 - Provide additional methods in cardputer_keyboard:
    - detach_serial()
    - attach_serial()
    - key_to_char(key, shift)
