--- ports/espressif/boards/m5stack_cardputer/board.h
@@ -1,9 +1,0 @@
-// This file is part of the CircuitPython project: https://circuitpython.org
-//
-// SPDX-FileCopyrightText: Copyright (c) 2020 Scott Shawcroft for Adafruit Industries
-//
-// SPDX-License-Identifier: MIT
-
-#include "shared-bindings/keypad_demux/DemuxKeyMatrix.h"
-
-extern keypad_demux_demuxkeymatrix_obj_t board_keyboard;
